restrict scale to chosen types
