+++ crates/db/src/kv/codecs/mod.rs
@@ -1,0 +1,1 @@
+mod scale;
